+++ vendor-cookbooks/samba/metadata.rb
@@ -1,0 +1,15 @@
+name             "samba"
+maintainer       "Opscode, Inc."
+maintainer_email "cookbooks@opscode.com"
+license          "Apache 2.0"
+description      "Installs/Configures samba"
+long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
+version          "0.11.4"
+
+recipe "samba::default", "Includes the samba::client recipe"
+recipe "samba::client", "Installs smbclient package"
+recipe "samba::server", "Installs samba server packages and configures smb.conf"
+
+%w{ arch debian ubuntu centos fedora redhat scientific amazon }.each do |os|
+  supports os
+end