+++ vendor-cookbooks/xml/metadata.rb
@@ -4,9 +4,10 @@ maintainer_email 'cookbooks@opscode.com'
 license          'Apache 2.0'
 description      'Installs xml'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '1.2.4'
+version          '1.2.6'
 
 depends 'build-essential'
+depends 'chef-sugar'
 
 supports 'amazon'
 supports 'arch'