+++ vendor-cookbooks/mysql/CHANGELOG.md
@@ -3,6 +3,21 @@ mysql Cookbook CHANGELOG
 This file is used to list changes made in each version of the mysql cookbook.
 
 
+v5.3.6 (2014-06-18)
+-------------------
+- Fixing pid path location. Updating tests to include real RHEL
+
+
+v5.3.4 (2014-06-16)
+-------------------
+- Fixing specs for Amazon Linux server package names
+
+
+v5.3.2 (2014-06-16)
+-------------------
+- Fixing Amazon Linux support
+
+
 v5.3.0 (2014-06-11)
 -------------------
 - #189 - Fix server_repl_password description